Added compatibility layer for SimpleEditableStyledDocument, removed storing segment operations in ReadOnlyStyledDocument
