+++ richtextfx/src/main/java/org/fxmisc/richtext/model/ReadOnlyStyledDocument.java
@@ -84,20 +84,20 @@ public final class ReadOnlyStyledDocument<PS, SEG, S> implements StyledDocument<
         String last = str.substring(start);
         res.add(new Paragraph<>(paragraphStyle, segmentOps, segmentOps.create(last, style)));
 
-        return new ReadOnlyStyledDocument<>(res, segmentOps);
+        return new ReadOnlyStyledDocument<>(res);
     }
 
     public static <PS, SEG, S> ReadOnlyStyledDocument<PS, SEG, S> fromSegment(SEG segment,  PS paragraphStyle, S style, SegmentOps<SEG, S> segmentOps) {
         Paragraph<PS, SEG, S> content = new Paragraph<PS, SEG, S>(paragraphStyle, segmentOps, Arrays.asList(segment));
         List<Paragraph<PS, SEG, S>> res = Arrays.asList(content);
-        return new ReadOnlyStyledDocument<>(res, segmentOps);
+        return new ReadOnlyStyledDocument<>(res);
     }
 
     public static <PS, SEG, S> ReadOnlyStyledDocument<PS, SEG, S> from(StyledDocument<PS, SEG, S> doc) {
         if(doc instanceof ReadOnlyStyledDocument) {
             return (ReadOnlyStyledDocument<PS, SEG, S>) doc;
         } else {
-            return new ReadOnlyStyledDocument<>(doc.getParagraphs(), doc.getSegOps());
+            return new ReadOnlyStyledDocument<>(doc.getParagraphs());
         }
     }
 
@@ -118,7 +118,7 @@ public final class ReadOnlyStyledDocument<PS, SEG, S> implements StyledDocument<
 
             @Override
             public StyledDocument<PS, SEG, S> decode(DataInputStream is) throws IOException {
-                return new ReadOnlyStyledDocument<>(codec.decode(is), segmentOps);
+                return new ReadOnlyStyledDocument<>(codec.decode(is));
             }
 
         };
@@ -154,21 +154,15 @@ public final class ReadOnlyStyledDocument<PS, SEG, S> implements StyledDocument<
     private String text = null;
     private List<Paragraph<PS, SEG, S>> paragraphs = null;
 
-    private final SegmentOps<SEG, S> segmentOps;
-    @Override
-    public final SegmentOps<SEG, S> getSegOps() { return segmentOps; }
-
-    private ReadOnlyStyledDocument(NonEmptyFingerTree<Paragraph<PS, SEG, S>, Summary> tree, SegmentOps<SEG, S> segmentOps) {
+    private ReadOnlyStyledDocument(NonEmptyFingerTree<Paragraph<PS, SEG, S>, Summary> tree) {
         this.tree = tree;
-        this.segmentOps = segmentOps;
     }
 
-    ReadOnlyStyledDocument(List<Paragraph<PS, SEG, S>> paragraphs, SegmentOps<SEG, S> segmentOps) {
+    ReadOnlyStyledDocument(List<Paragraph<PS, SEG, S>> paragraphs) {
         this.tree =
                 FingerTree.mkTree(paragraphs, summaryProvider()).caseEmpty().unify(
                         emptyTree -> { throw new AssertionError("Unreachable code"); },
                         neTree -> neTree);
-        this.segmentOps = segmentOps;
     }
 
     @Override
@@ -222,8 +216,8 @@ public final class ReadOnlyStyledDocument<PS, SEG, S> implements StyledDocument<
         return tree.splitAt(row).map((l, p, r) -> {
             Paragraph<PS, SEG, S> p1 = p.trim(col);
             Paragraph<PS, SEG, S> p2 = p.subSequence(col);
-            ReadOnlyStyledDocument<PS, SEG, S> doc1 = new ReadOnlyStyledDocument<>(l.append(p1), segmentOps);
-            ReadOnlyStyledDocument<PS, SEG, S> doc2 = new ReadOnlyStyledDocument<>(r.prepend(p2), segmentOps);
+            ReadOnlyStyledDocument<PS, SEG, S> doc1 = new ReadOnlyStyledDocument<>(l.append(p1));
+            ReadOnlyStyledDocument<PS, SEG, S> doc2 = new ReadOnlyStyledDocument<>(r.prepend(p2));
             return t(doc1, doc2);
         });
     }
@@ -246,7 +240,7 @@ public final class ReadOnlyStyledDocument<PS, SEG, S> implements StyledDocument<
         FingerTree<Paragraph<PS, SEG, S>, Summary> tree2 = (other instanceof ReadOnlyStyledDocument)
                 ? ((ReadOnlyStyledDocument<PS, SEG, S>) other).tree.split(1)._2
                 : FingerTree.mkTree(other.getParagraphs().subList(1, other.getParagraphs().size()), summaryProvider());
-        return new ReadOnlyStyledDocument<>(tree1.join(tree2), segmentOps);
+        return new ReadOnlyStyledDocument<>(tree1.join(tree2));
     }
 
     @Override
@@ -305,7 +299,7 @@ public final class ReadOnlyStyledDocument<PS, SEG, S> implements StyledDocument<
         for(int i = 0; i < n; ++i) {
             pars.add(f.apply(tree.getLeaf(i)));
         }
-        return new ReadOnlyStyledDocument<>(pars, segmentOps);
+        return new ReadOnlyStyledDocument<>(pars);
     }
 
     @Override
